Muevo funciones de tipos fuera de helpersW
