+++ ckanext/gobar_theme/helpers/__init__.py
@@ -3,3 +3,4 @@ from .helpers import *
 from .organizations import *
 from .groups import *
 from .licenses import *
+from .types import *
