+++ ckanext/gobar_theme/helpers/licenses.py
@@ -1,5 +1,4 @@
 #!coding=utf-8
-
 from ckan import model as model
